feat: refactor component imports to use new Util path for OntologyTerm, TimeElementDisplay, and QuantityDisplay
